ex04_dashboard/requirements.txt
streamlit==1.46.1
pandas==2.2.3
psycopg2-binary==2.9.10
